refactor and fixes path for elementor widget
