+++ package.json
@@ -1,36 +1,36 @@
 {
-	"name": "reduct-video-plugin",
-	"version": "1.3.0",
-	"description": "",
-	"main": "src/index.js",
-	"scripts": {
-		"start": "./node_modules/.bin/wp-scripts start",
-		"build": "./node_modules/.bin/wp-scripts build",
-		"check-engines": "./node_modules/.bin/wp-scripts check-engines",
-		"check-licenses": "./node_modules/.bin/wp-scripts check-licenses",
-		"format": "./node_modules/.bin/wp-scripts format",
-		"lint:css": "./node_modules/.bin/wp-scripts lint-style",
-		"lint:js": "./node_modules/.bin/wp-scripts lint-js",
-		"lint:md:docs": "./node_modules/.bin/wp-scripts lint-md-docs",
-		"lint:pkg-json": "./node_modules/.bin/wp-scripts lint-pkg-json",
-		"packages-update": "./node_modules/.bin/wp-scripts packages-update",
-		"bundle": "yarn build && ./node_modules/.bin/wp-scripts plugin-zip"
-	},
-	"keywords": [],
-	"author": "Reduct Video",
-	"license": "ISC",
-	"files": [
-		"reduct-video-plugin.php",
-		"reduct-embed-elementor.php",
-		"reduct-embed-elementor-widget.php",
-		"template.php",
-		"build/index.js",
-		"build/index.asset.php",
-		"build/elementorWidget.js",
-		".npmignore"
-	],
-	"devDependencies": {
-		"@wordpress/scripts": "^26.2.0",
-		"terser-webpack-plugin": "^5.3.7"
-	}
+  "name": "reduct-video-plugin",
+  "version": "1.3.0",
+  "description": "",
+  "main": "src/index.js",
+  "scripts": {
+    "start": "./node_modules/.bin/wp-scripts start",
+    "build": "./node_modules/.bin/wp-scripts build",
+    "check-engines": "./node_modules/.bin/wp-scripts check-engines",
+    "check-licenses": "./node_modules/.bin/wp-scripts check-licenses",
+    "format": "./node_modules/.bin/wp-scripts format",
+    "lint:css": "./node_modules/.bin/wp-scripts lint-style",
+    "lint:js": "./node_modules/.bin/wp-scripts lint-js",
+    "lint:md:docs": "./node_modules/.bin/wp-scripts lint-md-docs",
+    "lint:pkg-json": "./node_modules/.bin/wp-scripts lint-pkg-json",
+    "packages-update": "./node_modules/.bin/wp-scripts packages-update",
+    "bundle": "yarn build && ./node_modules/.bin/wp-scripts plugin-zip"
+  },
+  "keywords": [],
+  "author": "Reduct Video",
+  "license": "ISC",
+  "files": [
+    ".npmignore",
+    "build/index.js",
+    "build/index.asset.php",
+    "build/elementorWidget.js",
+    "reduct-video-plugin.php",
+    "template.php",
+    "widget/emptyTranscript.png",
+    "widget/reduct-embed-elementor-widget.php"
+  ],
+  "devDependencies": {
+    "@wordpress/scripts": "^26.2.0",
+    "terser-webpack-plugin": "^5.3.7"
+  }
 }
