+++ reduct-video-plugin.php
@@ -11,11 +11,17 @@ if (!defined('ABSPATH')) // exit if try to access from the browser directly
 
 define("VIDEO_RESOURCE_URL", "https://app.reduct.video/e/");
 
-class Plugin 
+class Plugin
 {
     function __construct()
     {
         add_action('init', array($this, 'adminAssets'));
+
+        // register elementor widget
+        add_action('elementor/widgets/register', array($this, 'register_reduct_reel_embed_widget'));
+        add_action('elementor/frontend/after_enqueue_scripts', array($this, 'enqueue_custom_script'));
+
+        // register routes
         add_action('rest_api_init', array($this, 'rest_api_routes'));
     }
 
@@ -43,6 +49,27 @@ class Plugin
         $output = ob_get_clean();
         return $output;
     }
+
+    function register_reduct_reel_embed_widget($widgets_manager)
+    {
+
+        require_once(__DIR__ . '/widget/reduct-embed-elementor-widget.php');
+
+        $widgets_manager->register(new \Elementor_Reduct_Reel_Embed_Widget());
+    }
+
+    function enqueue_custom_script()
+{
+	wp_enqueue_script(
+		'elementorWidget',
+		plugin_dir_url(__FILE__) . 'build/elementorWidget.js',
+		array('jquery','wp-blocks', 'wp-element', 'wp-components'),
+		'1.0.0',
+		true
+	);
+
+	wp_localize_script('elementorWidget', 'WP_PROPS', array('site_url' => get_site_url()));
+}
 
     function transcript_route($request)
     {
